pushing to github

+++ tranx.go
@@ -3,7 +3,7 @@ package chipmunk
 import "time"
 
 type tranx struct {
-	Cost  float32    `json:"cost"`
+	Cost  float64    `json:"cost"`
 	Store string     `json:"store"`
 	Info  string     `json:"Info"`
 	Month time.Month `json:"Month"`
